fix page selection
