+++ packages/types/src/message/common/index.ts
@@ -2,5 +2,6 @@ export * from './base';
 export * from './image';
 export * from './messageGroup';
 export * from './metadata';
+export * from './pageSelection';
 export * from './tools';
 export * from './translate';
