+++ src/features/Conversation/Messages/User/components/MessageContent.tsx
@@ -2,12 +2,12 @@ import { Flexbox } from '@lobehub/ui';
 import { memo } from 'react';
 
 import MarkdownMessage from '@/features/Conversation/Markdown';
-import PageSelections from '@/features/Conversation/Messages/components/Extras/PageSelections';
 import { type UIChatMessage } from '@/types/index';
 
 import { useMarkdown } from '../useMarkdown';
 import FileListViewer from './FileListViewer';
 import ImageFileListViewer from './ImageFileListViewer';
+import PageSelections from './PageSelections';
 import VideoFileListViewer from './VideoFileListViewer';
 
 const UserMessageContent = memo<UIChatMessage>(
